Airflow Ingestion Pipeline/requirements.txt
aext-assistant
aext-assistant-server
aext-core
aext-core-server
aext-panels
aext-panels-server
aext-project-filebrowser-server
aext-share-notebook
aext-share-notebook-server
aext-shared
aext-toolbox
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alabaster
alembic
altair
anaconda-anon-usage
anaconda-catalogs
anaconda-cli-base
anaconda-client
anaconda-cloud-auth
anaconda-navigator
anaconda-project
annotated-types
anyio
appdirs==1.4.4
applaunchservices
appnope==0.1.4
appscript
archspec
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
astropy-iers-data
asttokens==3.0.0
async-lru
atomicwrites==1.4.0
attrs
Automat
autopep8
Babel
bcrypt
beautifulsoup4
binaryornot
black
bleach
blinker
bokeh
boltons
botocore
Bottleneck
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
colorcet
comm==0.2.2
conda
conda-build
conda-content-trust
conda-libmamba-solver
conda-pack
conda-package-handling
conda-repo-cli
conda-token
conda_index
conda_package_streaming
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cytoolz
dask
dask-expr
datashader
debugpy==1.8.14
decorator==5.2.1
defusedxml
diff-match-patch
dill
distributed
distro
dmglib
docstring-to-markdown
docutils
et-xmlfile
executing==2.2.0
fastjsonschema
filelock
flake8
Flask
fonttools
frozendict
frozenlist
fsspec
gensim
gitdb
GitPython
greenlet
h11
h5py
HeapDict
holoviews
httpcore
httpx
hvplot
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
incremental
inflection
iniconfig
intake
intervaltree
ipykernel==6.29.5
ipython==9.2.0
ipython-genutils
ipython_pygments_lexers==1.1.1
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jaraco.classes
jedi==0.19.2
jellyfish
Jinja2
jmespath
joblib
json5
jsonpatch
jsonpointer==2.1
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keyring
kiwisolver
lazy-object-proxy
lazy_loader
lckr_jupyterlab_variableinspector
libarchive-c
libmambapy
linkify-it-py
llvmlite
lmdb
locket
lxml
lz4
Mako
Markdown
markdown-it-py
MarkupSafe
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe
mdit-py-plugins
mdurl
menuinst
mistune
more-itertools
mpmath
msgpack
multidict
multipledispatch
mypy
mypy-extensions
navigator-updater
nb_conda_kernels
nbclient
nbconvert
nbformat
nest-asyncio==1.6.0
networkx
nltk
notebook
notebook_shim
numba
numexpr
numpy
numpydoc
openpyxl
overrides
packaging
pandas
pandocfilters
panel
param
parsel
parso==0.8.4
partd
pathspec
patsy
pexpect==4.9.0
pickleshare
pillow
pkce
pkginfo
platformdirs==4.3.8
plotly
pluggy
ply
prometheus-client
prompt_toolkit==3.0.51
Protego
protobuf==4.25.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules==0.2.8
pycodestyle
pycosat
pycparser
pyct
pycurl
pydantic
pydantic-settings
pydantic_core
pydeck
PyDispatcher
pydocstyle
pyerfa
pyflakes
Pygments==2.19.1
PyJWT
pylint
pylint-venv
pyls-spyder==0.4.0
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-CoreServices
pyobjc-framework-FSEvents
pyodbc
pyOpenSSL
pyparsing
PyQt5==5.15.10
PyQt5-sip
PyQtWebEngine==5.15.6
PySocks
pytest
python-dateutil==2.9.0.post0
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-server
python-slugify
pytoolconfig
pytz
pyviz_comms
PyWavelets
PyYAML
pyzmq==26.4.0
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
readchar
referencing
regex
requests
requests-file
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rope
rpds-py
Rtree
ruamel-yaml-conda
ruamel.yaml
ruamel.yaml.clib
s3fs
scikit-image
scikit-learn
scipy
Scrapy
seaborn
semver
Send2Trash
sentry-sdk
service-identity
setuptools==75.1.0
shellingham
sip
six==1.17.0
smart-open
smmap
sniffio
snowballstemmer
sortedcontainers
soupsieve
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
stack-data==0.6.3
statsmodels
streamlit
sympy
tables
tabulate
tblib
tenacity
terminado
text-unidecode
textdistance
threadpoolctl
three-merge
tifffile
tinycss2
tldextract
toml
tomli
tomlkit
toolz
tornado==6.5.1
tqdm
traitlets==5.14.3
truststore
Twisted
typer
typing_extensions
tzdata
uc-micro-py
ujson
unicodedata2
Unidecode
urllib3
w3lib
watchdog
wcwidth==0.2.13
webencodings
websocket-client
Werkzeug
whatthepatch
wheel==0.44.0
widgetsnbextension
wrapt
wurlitzer
xarray
xlwings
xyzservices
yapf
yarl
zict
zipp
zope.interface
zstandard
langchain_text_splitters
edgartools
langchain
qdrant-client
datasets
langchain_community
sentence-transformers
peft
transformers
